+++ OpenLayersWebMap/algorithm1.js
@@ -1,0 +1,5 @@
+function implementAlgorithm() {
+  console.log("hello there");
+}
+
+window.implementAlgorithm = implementAlgorithm;
