+++ OpenLayersWebMap/index.html
@@ -79,6 +79,10 @@
 	<!-- Load your modules -->
 	<script type="module" src="graphConverter.js"></script>
 	<script type="module" src="graphologyConverter.js"></script>
+	<script type="module" src="services/GpxServices.js"></script>
+	<script type="module" src="services/OverpassServices.js"></script>
+	<script type="module" src="utils/GraphFunctions.js"></script>
+	<script type="module" src="Algorithm1.js"></script>
 	<script src="Edge.js"></script>
 	<script src="Node1.js"></script>
 	<script src="Route.js"></script>
